pyproject.toml
[project]
name = "tasktree"
version = "0.1.0"
description = "A task automation tool with intelligent incremental execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "colorama>=0.4.6",
    "pygments>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
tt = "tasktree.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
